+++ Dyalect/Linker/Lang.cs
@@ -58,7 +58,7 @@ namespace Dyalect.Linker
         }
 
         [Function("makeArray")]
-        private DyObject makeArray(ExecutionContext ctx, DyObject[] args)
+        public DyObject MakeArray(ExecutionContext ctx, DyObject[] args)
         {
             var size = args[0];
             var n = size.GetInteger();
@@ -88,6 +88,22 @@ namespace Dyalect.Linker
                     la1 != null ? la1.Value : a1,
                     a2.TypeId == StandardType.Label ? (la2 = (DyLabel)a2).Label : null,
                     la2 != null ? la2.Value : a2
+                    );
+            }
+
+            if (len == 3)
+            {
+                var a1 = args[0];
+                var a2 = args[1];
+                var a3 = args[2];
+                DyLabel la1 = null, la2 = null, la3 = null;
+                return DyTuple.Create(
+                    a1.TypeId == StandardType.Label ? (la1 = (DyLabel)a1).Label : null,
+                    la1 != null ? la1.Value : a1,
+                    a2.TypeId == StandardType.Label ? (la2 = (DyLabel)a2).Label : null,
+                    la2 != null ? la2.Value : a2,
+                    a3.TypeId == StandardType.Label ? (la3 = (DyLabel)a3).Label : null,
+                    la3 != null ? la3.Value : a3
                     );
             }