+++ Dyalect/Runtime/Types/DyString.cs
@@ -109,10 +109,7 @@ namespace Dyalect.Runtime.Types
         protected override DyObject LengthOp(DyObject arg, ExecutionContext ctx)
         {
             var len = arg.GetString().Length;
-            return len == 0 ? DyInteger.Zero
-                : len == 1 ? DyInteger.One
-                : len == 2 ? DyInteger.Two
-                : new DyInteger(len);
+            return DyInteger.Get(len);
         }
 
         protected override DyString ToStringOp(DyObject arg, ExecutionContext ctx) => StringUtil.Escape(arg.GetString());